remove di from tests
