+++ src/main/kotlin/burrows/apps/example/gif/presentation/di/module/RiffsyModule.kt
@@ -1,5 +1,6 @@
 package burrows.apps.example.gif.presentation.di.module
 
+import burrows.apps.example.gif.BuildConfig
 import burrows.apps.example.gif.data.rest.repository.RiffsyApiClient
 import burrows.apps.example.gif.presentation.di.scope.PerActivity
 import dagger.Module
@@ -10,10 +11,10 @@ import retrofit2.Retrofit
  * @author [Jared Burrows](mailto:jaredsburrows@gmail.com)
  */
 @Module
-open class RiffsyModule {
-  @Provides @PerActivity open fun providesRiffsyApi(retrofit: Retrofit.Builder): RiffsyApiClient {
+class RiffsyModule {
+  @Provides @PerActivity fun providesRiffsyApi(retrofit: Retrofit.Builder): RiffsyApiClient {
     return retrofit
-      .baseUrl(RiffsyApiClient.BASE_URL)
+      .baseUrl(BuildConfig.BASE_URL)
       .build()
       .create(RiffsyApiClient::class.java)
   }
