+++ src/test/resources/search_results.json
@@ -13,7 +13,7 @@
               138
             ],
             "duration": 2.0,
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           },
           "nanowebm": {
             "url": "https://media.riffsy.com/videos/9e54e727e3f615c52f611ce94ccfa400/webm",
@@ -21,15 +21,15 @@
               150,
               138
             ],
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           },
           "tinygif": {
-            "url": "https://media.riffsy.com/images/6088f94e6eb5dd7584dedda0fe1e52e1/raw",
+            "url": "http://localhost:8080/images/6088f94e6eb5dd7584dedda0fe1e52e1/raw",
             "dims": [
               220,
               203
             ],
-            "preview": "https://media.riffsy.com/images/6f2ed339fbdb5c1270e29945ee1f0d77/raw",
+            "preview": "http://localhost:8080/images/6f2ed339fbdb5c1270e29945ee1f0d77/raw",
             "size": 18228
           },
           "tinymp4": {
@@ -39,7 +39,7 @@
               296
             ],
             "duration": 2.0,
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           },
           "tinywebm": {
             "url": "https://media.riffsy.com/videos/b69f65ebe587be24c2786f264dde63d3/webm",
@@ -47,7 +47,7 @@
               320,
               296
             ],
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           },
           "webm": {
             "url": "https://media.riffsy.com/videos/01a8443eb106065b258391878c9f326b/webm",
@@ -55,15 +55,15 @@
               540,
               500
             ],
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           },
           "gif": {
-            "url": "https://media.riffsy.com/images/5b6a39aa00312575583031d2de4edbd4/raw",
+            "url": "http://localhost:8080/images/5b6a39aa00312575583031d2de4edbd4/raw",
             "dims": [
               498,
               461
             ],
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw",
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw",
             "size": 985952
           },
           "mp4": {
@@ -73,15 +73,15 @@
               500
             ],
             "duration": 2.0,
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           },
           "nanogif": {
-            "url": "https://media.riffsy.com/images/f85bbd038a243b1a21ca3fe8667636dd/raw",
+            "url": "http://localhost:8080/images/f85bbd038a243b1a21ca3fe8667636dd/raw",
             "dims": [
               97,
               90
             ],
-            "preview": "https://media.riffsy.com/images/b583e30155de53bbcd43ec91b24f1f1b/raw",
+            "preview": "http://localhost:8080/images/b583e30155de53bbcd43ec91b24f1f1b/raw",
             "size": 8338
           },
           "loopedmp4": {
@@ -91,7 +91,7 @@
               500
             ],
             "duration": 6.0,
-            "preview": "https://media.riffsy.com/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
+            "preview": "http://localhost:8080/images/21accb6fe0c3d2ab562b9d429c7af88c/raw"
           }
         }
       ],
